requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
greenlet==3.1.1
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==9.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
